% Generated by roxygen2 (4.0.2): do not edit by hand
\name{custom_peer}
\alias{custom_peer}
\title{Run PEER}
\usage{
custom_peer(expr, n.factors = NULL, covars = NULL)
}
\arguments{
\item{expr}{A gene expression matrix with dimensions N x g.}

\item{n.factors}{Number of factors to be estimated, if left unspecified 25% of the sample size
       is going to be used. (round(N * 0.25))}

\item{covars}{Known fixed effects that you wish to account for.}
}
\value{
List with residuals, weights and factors saved.
}
\description{
Custom PEER function that generates factor coefficients and residuals.
}
\keyword{keywords}

