% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eqtl_simulator}
\alias{eqtl_simulator}
\title{Simulating eQTL phenotypes.}
\usage{
eqtl_simulator(genotypes, n.pheno = NULL, n.eqtl = NULL,
  cis.trans.ratio = NULL, trans.impact = NULL, trans.nerf = NULL,
  coeff.sample = NULL, cis.sd = NULL)
}
\arguments{
\item{genotypes}{A genotype matrix with dimensions N x g.}

\item{n.pheno}{Number of phenotypes to simulate}

\item{n.eqtl}{Number of eQTL pairs}

\item{cis.trans.ratio}{Ratio between cis and trans hits}

\item{trans.impact}{Maximum percentage of total genes a trans hit can influence.}

\item{trans.nerf}{Ratio of trans effect size compared to cis-hits.
Accounts for the observation that trans hits usually have smaller
effect sizes.}

\item{coeff.sample}{A vector from which the effect sizes (coefficiens) are
       going to be sampled from.}
}
\value{
Phenotype matrix with dimensions N x n.pheno
}
\description{
Generates a phenotype matrix based on a given input genotype matrix.
}
\keyword{keywords}

