% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eqtl_simulator_cistrans_h5}
\alias{eqtl_simulator_cistrans_h5}
\title{Simulating eQTL phenotypes.}
\usage{
eqtl_simulator_cistrans_h5(input_h5 = NULL, n_pheno = 2000, n_geno = NULL,
  n_eqtl = 1500, cis_trans_ratio = 0.7, simulation.id = "sim1",
  output.path = "./", coeff_mean = 2.5, trans_nerf = 0.7, hot_spots = 8,
  hot_spot_impact = 0.1, hidden.factor = TRUE, factors = "sparse",
  factor_coeff = 2.1, hf_frac = 0.2, effect_size = 2,
  factor_type = "sparse")
}
\arguments{
\item{genotypes}{A genotype matrix with dimensions N x g.}

\item{n.pheno}{Number of phenotypes to simulate}

\item{n.eqtl}{Number of eQTL pairs}

\item{cis.trans.ratio}{Ratio between cis and trans hits}

\item{trans.impact}{Maximum percentage of total genes a trans hit can influence.}

\item{trans.nerf}{Ratio of trans effect size compared to cis-hits.
Accounts for the observation that trans hits usually have smaller
effect sizes.}

\item{coeff.sample}{A vector from which the effect sizes (coefficiens) are
       going to be sampled from.}
}
\value{
Phenotype matrix with dimensions N x n.pheno
}
\description{
Generates a phenotype matrix based on a given input genotype matrix.
}
\keyword{keywords}

