% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eqtl_simulator_h5}
\alias{eqtl_simulator_h5}
\title{Simulating eQTL phenotypes.}
\usage{
eqtl_simulator_h5(input.geno = NULL, n.pheno, n.eqtl, simulation.id,
  output.path = "./simpheno", coeff.mean = 2.5, n.sim = 5,
  cis.trans.ratio = 0.7, trans.impact = 0.1, trans.nerf = 0.7,
  hidden.factor = TRUE, factors = NULL, factor.coeff = NULL,
  hf.frac = 0.2, effect.size = 2)
}
\arguments{
\item{n.pheno}{Number of phenotypes to simulate}

\item{n.eqtl}{Number of eQTL pairs}

\item{cis.trans.ratio}{Ratio between cis and trans hits}

\item{trans.impact}{Maximum percentage of total genes a trans hit can influence.}

\item{trans.nerf}{Ratio of trans effect size compared to cis-hits.
Accounts for the observation that trans hits usually have smaller
effect sizes.}

\item{genotypes}{A genotype matrix with dimensions N x g.}

\item{coeff.sample}{A vector from which the effect sizes (coefficiens) are
       going to be sampled from.}
}
\value{
Phenotype matrix with dimensions N x n.pheno
}
\description{
Generates a phenotype matrix based on a given input genotype matrix.
}
\keyword{keywords}

