% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hf_sim}
\alias{hf_sim}
\title{Simulating hidden factor effects}
\usage{
hf_sim(n.genes, n.samples, hf.type = c("sparse", "normal", "uniform"),
  coeff.dist = c("normal", "uniform", "binom"), fraction.affected = 0.2,
  binom.prob = 0.4, factor.effect.size = 2)
}
\arguments{
\item{n.genes}{Number of genes in the original phenotype matrix.}

\item{n.samples}{Number of samples in the original phenotype matrix}

\item{fraction.affected}{Fraction of genes affected by the hidden factor.}
}
\value{
hidden factor matrix with dimensions n.samples x n.genes
}
\description{
Generates a phenotype matrix based on a given input genotype matrix.
}
\keyword{keywords}

