% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pheno_position_simulator}
\alias{pheno_position_simulator}
\title{Simulating phenotype position information}
\usage{
pheno_position_simulator(n.pheno, n.chr)
}
\arguments{
\item{n.pheno}{Number of phenotypes to use}

\item{n.chr}{Number of chromosomes to simulate}
}
\value{
Dataframe with phenotype, pheno_chr, pheno_start, pheno_end, idx as columns
}
\description{
Generates a data frame with simulated chromosome position information
}

